Type aliases specifying a type variable to a concrete type now translates (more) correctly.
Previously, things roughly like `type alias JavaBool = Maybe Bool` resulted in type signatures referring to `Maybe a` where `JavaBool` was used. Spotted by Mario because no `ElmVal' a` annotation was generated, which was actually correct behaviour, but it didn't fit with `Maybe a` seen in the rest of the type signature.